// packages/representante/src/types/TConsumer.ts
export type TConsumer = {
  consumer_balance: number
  consumer_birthDate: string
  consumer_blockedBalance: number
  consumer_cpf: string
  consumer_createdAt: string
  consumer_deactivedAccount: boolean
  consumer_email: string
  consumer_emailConfirmated: boolean
  consumer_fullName: string
  consumer_id: string
  consumer_phone?: string
  consumer_signatureRegistered: boolean
  address_complement?: string
  address_district?: string
  address_number?: number
  address_street?: string
  city_name: string
  address_zipCode: string
}